Commented out lines 212-214 per this email from Eric:
From briefly looking at it it looks like there's this bit of code in the populateShows function that skips the first result when loading more shows via the infinite scroll. This is in a bit of code that loops over each show in the json returned by geteventsbetween. From a quick glance, it looks like we could just remove it, since the comment seems to be lying when it says that the show is already displayed.